Add delete_attachments to schema
DummyDefinition being in the same file than schema test caused
mock library to fail in some cases (as it documentation says,
because of the way mock overrides the module, it must be defined
in a separate file from the test file). That's why definition
module was declared in setup for each test, causing some warnings
that are gone now.